fix small thing
