+++ winterfell/controllers/claimController.js
@@ -1,17 +1,17 @@
 var _ = require('lodash');
-var ApiLog = require('./../middlewares/api-logger');
 var auth = require('../middlewares/auth');
 var Claim = require('../models/claim');
 var ClaimService = require('./../services/claimService');
 var Form = require('../models/form');
 var http = require('http-status-codes');
 var httpErrors = require('./../utils/httpErrors');
+var LogHelper = require('./../utils/logHelper');
 var mongoose = require('mongoose');
 var User = require('../models/user');
 var UserValues = require('../models/userValues');
 
 module.exports = function (app) {
-  var middlewares = [auth.authenticatedOr404, ApiLog.logApi];
+  var middlewares = [auth.authenticatedOr404, LogHelper.logApi];
 
   // Get all claims for a user
   app.get('/claims', middlewares, function (req, res) {
