+++ winterfell/followLogs.sh
@@ -1,0 +1,2 @@
+tail -F ../logs/app.js | ./node_modules/.bin/bunyan
+
